Add terrain calculations
They are not integrated, but the calculations for
the terrain modifiers are there. The idea is that
you can have multiple modifiers from terrain, and those
will go into the full list for each unit, whereby
each unit will have the max and min selected for
them per rule 0.5b.