Add in boosting infantry, 10K simulations
Infantry class units will be boosted by artillery.
Next up is handling units that shoot multiple times better
and only participate in the first round. Also damaged
battleships are on the docket, and being able to pick
between the round types that were added a while back.
Also upped the simulation count to 10K because that
produces results that have very little variance.