Add swap, fix styling
This adds a button to quickly swap between attack
and defend so you can check the same battle from
both sides. Also fix the styling in both standard
and mobile modes so that it looks decent and doesn't
dance around when you swap stuff. Should be
pretty decent for that now.